rewriters: Remove exhR memoization, minor tweaks
This whole rewriter framework is very temporary. The exhR thing appears
to have been "bugged". I was getting extremely weird bug all the way up
in UIR that I was able to track down to an exhR cache load. Removing
exhR cache shouldn't hurt too much at the moment, because everything is
already too slow! Although with the new UIR, I did hit ~10x slowdown
without it. Which I obviously fixed by wrapping the UIR's exhR with
memoR, which is actually what one should do. In any case, as I've said,
this whole baz5 thing is a bunch of junk that I can't find enough time
to rewrite. We need a zipper based, term based rewriter. It's probably
going to be even slower but at least it's going to be correct.