uir: Upload the new implementation of UIR, tests, and resources
This new implementation requires at the very least some polishing of its
WwML. The UIs it can render are arguably on another level vs. what we
had before. We're getting closer and closer to browsers (although the
graphics & performance are so shitty I'm kind of wishing I haven't said
this). Let's say, in terms of the "palette of functionality".
Regardless, though, this impl, although a result of 8 iterations (and
lots of thinking), is still not "battle-tested" in the strongest sense.
Nor is it fast, again. It's very slow. It renders the test (which maybe
contains ~500 nodes?) in 1s. One, freaking, second. The test is worth
maybe 1ms at most. So we're talking 1000x slower than expected. I really
have no idea what to do about it. Yes there are lots of missing indexing
opportunities and low hanging fruit and opt isn't the focus right now.
But the low hanging fruit will end at some point. At 1000x, though? I
doubt it.