ml: Change the style of `SyntaxError` slightly
- Use ANSI colors. They are presumably theme-able, although I haven't
tested that out.
- Use undercurl for additional error highlight. I presume most modern
terminals support it. Those that do not, are hopefully going to ignore
it. I do not particularly like the undercurl that e.g. Alacritty draws
-- mostly because I cannot control its color separately from the text.
But it is closer to the point of highlighting an error anyway (error
undercurls must be red!! but I don't want the text itself to be red,
that's too distracting!)