ml: Upload the remainder of the huge reader update 🎉
It's impossible to split this into smaller commits so here it is,
another megacommit of mine.

We have a much more powerful reader now. Much more code, too -- and
nasty code that is. But as anything else practical, it got too nasty too
soon. I tried parser combinators but they were as nasty if not more,
esp. with my tendency to overcomplicate and overengineer things. Here,
with hand-written parsers and lexers, at least nastiness is expected and
not frowned upon. While exploring real world lexers and parsers of D,
Crystal, Golang and others -- gosh how nasty they are! We're in good
company!

NOTE: The WwML spec is currently moderately out of sync with this. Some
names are wrong in the spec and some things underdescribed or not
described at all. I hope I will find some time to fix these. There are
so many *more interesting* things to work on than a spec...