uir: Implement `x/y-branch` and improve layout messaging; fixes
- x/y-branch let us do responsive stuff like CSS container queries.
However, we *really* need Microfold to emit and keep track of them.
They're nasty to work with by hand.
- I've split the horizontal messaging flow into three stages, mail/in,
mail/mid, and mail/out. This lets us filter, ignore, or alter layout
messages depending on the node.
- This is currently used to implement the `may-float` hint, which lets a
floating node conditionally opt out of fr layout. The solution smells
very wrong, but this is the best thing that's passing the tests as of
now. I will try to come up with a fully local solution.