uir: Implement `x/y-signal` to trigger `sm` in enclosing `x/y-branch`
This is more general than `may-float`, while still passing all tests --
neat.

With x/y-branch (mainly), there is the nasty question of relayout when
the parent "rethinks" its dn-w/h, especially if it is able to
accommodate the `lg` variant. However, I'm currently shelving this since
I'm unable to come up with a test that would trigger this. The whole
thing seems to crumble under parent-driven relayout, but maybe it's me
thinking wrong.