ufold: Upload the new Microfold rewrite :tada:
(Subjectively) lots of engineering here and hopefully -- despite my
inexperience in the matter -- it turns out to be practical enough for
general use. Soma will be the primary user, though. User-facing
debuggability is a TODO. The machinery is there, but needs a lot more
work. We have clear passes. The part that I expect debugging be
necessary for is flow; which uses rewriters. Thus, first, rewriter
circuits will have to be implemented with proper debugging support. Then
we'll get flow debugging for free, effectively letting the user "step
through time" and see how styles are applied, just like it would be
possible with layout. All interactions become completely transparent to
anyone willing to look under the hood. There is practically no hidden
state.

A common interface exposing a Microfold&UIR&DwUIR environment is in
development, and hopefully will be available soon. This will allow us to
get rid of the nasty and heavily duplicated vdwuir, dwuir2ppm, and
uir2ppm tools.