m1: Add support for private/interior captures in `*All` operators
This is useful when we want to equate env dicts but not interior
captures (i.e. capture names must be the same, but we're not meaning to
equate across `%items` boundary). Although the place where this should
have been used is bugged out otherwise so this feature would be
unreachable for some time.