ww: Simplify `Ww::Issue` design
Spots are now unaware of their presentation, and are more or less
instructions for some kind of "presentation" or "spot printer" machine.
There's no such machine at the moment so the best we've got is Crystal's
built-in `pp`. The previous `to_omega` impl was too bloated, and it
doesn't belong to Spot/Issue. Moreover, Omega would probably move to a
term-based repr so I would have to rewrite this anyway.