ww: progress dump: New Alloy impl, improve rack windowing etc.
I'm tired of splitting these into small commits that nobody will ever
see, so here we go: another one of my mega-commits.

- General fixing, renaming, etc. of rack windowing machinery. Now under
the module WM.
- New Alloy impl. Very nice, because we're relying on Alloy very much at
this point, and the old extremely buggy impl became a roadblock & a huge
headache to work with. Moreover, the old impl was one of the few sources
of complete and utter crashes; mostly infinite recursion errors and
friends. The current thing is *vastly* more stable as it relies on the
most tested parts of the system.
- Alloy is tested separately as well, under the new `alloy.test.wwml`