ufold: Change the order of cascade to "first all presets, then all styles" of ancestors
Previously we were doing both simultaneously while descending through
ancestors. This messes up with my intuitive idea for how this should
work; on the other hand, the way it is done now, first presets of all
ancestors, then styles of all ancestors -- behaves just as I would
expect. This is the kind of thing that should remain invisible &
unknown-about if it works correctly. Now this appears to be the case.