ww: progress dump: Fixes, improved terminal rendering, extend Alloy
There's a lot of updates in here, some other ones are:

- `floating` supports `x` and `y` now, just like content; which means we
can make somewhat more complex layouts where a node leaves layout in one
axis but participates in another

- Simplify the whole require SDL / window implementations thing. Just
require all of them and let Crystal figure out what's used and what's
not. I'm tired of all this flag / mandatory require nonsense. The
project is monstrous anyway (or is slowly becoming monstrous, depending
on what "monstrous" means for you); a few extra requires shouldn't be a
problem.

- Begin Term Morph API. I hope it will replace all those
`morph`/`where`/`follow` methods on Dict, especially in light of the
upcoming dict rewrite (it won't be soon though, maybe in a few months;
the project is again in a state where there's no good user interface to
it; at least now there's Rack; but optimization is not a concern at the
moment).

- Fix typos and other minor changes.