ww: Upload `ww` tool and the new rack TUI frontend
Now there is a semi-visual interface to almost all of Wirewright; only
Meridium is not yet covered by racks, and the rest of stuff isn't
written yet. D7 will be rewritten and will be covered with racks; same
for pretty printing etc. I hope racks will be the unifying glue I wanted
so much. They are now, but whether they will stand the test of time --
we'll see.