ww: Fix (hopefully!) overflow errors in event log
The reason was, as far as I understand, that viewport passed height
through, which it should have, of course; it's an opt out. On the other
hand, isn't the goal of a viewport to *not* pass height through by
default, since viewports exist to clip, and thus allow infinite
width/height/both within? Anyway. Using `y-content` to restrict height
flow past the viewport itself appears to solve the issue. Now we clip
something that's arbitrarily high, which is expected behavior;
previously, viewport contents tried to cram themselves into the limited
height of the viewport itself, which is expected to be glitchy; it's the
layout alternative of UB, and I'm fine with it; not happy about it, but
fine.