term: Add a null check to prevent Crystal bugs from segfaulting us
There is a Crystal bug (apparently a codegen bug of some sort) that I
fail to reproduce for certain, related to procs and return types of
procs. It sometimes causes bizarre crashes that have the smell of wrong
type being (generated for) for values returned from procs. I have
learned to fix it by feel when it happens; but a check like this would
help to pinpoint it anyway.