rack: Add experimental `dwuir/console` node & support it in rack TUI
Rack TUI now acts as a "multiplexer" for `dwuir/console`s, showing them
as tabs along with the rack interface itself. The whole thing is
partially working, although we're missing replier support with the uir
node (I'm considering to rename repliers to rulers because that's more
functionally correct). So the only remaining problem, if we forget
performance, which is utterly horrible, especially Microfold... as I was
saying, the only remaining problem is in adding support for proper size
reports from tDwUIR's ruler rather than the current gDwUIR. So I think
after we have `ruler/textual` and `ruler/graphical` nodes, and after
`uir` can talk to them, we'll have decent GUI & TUI support. Then
there's the last frontier, implementing the rack TUI with rack itself.
After that's done I'll be more or less convinced that racks work as an
idea.