rack: Add support for different repliers (metrics) in `uir` node
This is a hack on top of a hack. uiR and metricsR should be different
rewriters. uiR shouldn't be a special case node. Whatever. This works
and will work for now. My main focus should be on getting D7 to draw
again, implement rack with rack to (im)prove the approach (until it)
works, and then some pattern refactors & opt. Then maybe I'll do
rewriter circuits because it's a huge missing piece in the sense of lots
of related TODOs. Then we'll be able to work on improving things like
these.