rack: Add support for sizes edge in `dwuir/console`
The sizes edge now reports correct viewport sizes on rack activate and
viewport size change. This is achieved through observing a sized node's
width and height in output of uiR just before it goes to DwUIR; and
sending those width and height to the server. The rack can then depend
on the knowledge of size before producing the window to render. This
seems to be a solution with the least communication overhead.

With Microfold, we have a new property `sized-id`, which goes to
`sized-box` under `id`.