rack: Moderate refactors, move to SDL wait-based event loop
We are now using SDL user events instead of the previous abomination of
a main loop. This triggered a chain reaction of refactors in related
areas. Everything appears clean now, at least cleaner than before. We're
not polling needlessly now, nor are we doing anything framerate-y or
game loop-ey on the server side. This refactor, besides cleaning lots of
stuff up, allows us to proceed into fixpoint territory, which is
required for further progress on racks. That is, we want to allow
feedback loops run through a device, maybe named `fixpoint`, that will
make sure to do rewriting as well as event handling. This means its core
must reside in the event loop, which was borked before this refactor.

There are also two new rewriters: `uiR` and `metricsR/*`. It's only a
matter of wiring them up through `fixpoint` now. D7 nodes will work
similarly.