rack: Add `d7-editR/⸨repair,post⸩`, `gate` nodes, bugfixes
I have already tried to build a rack that is like the very first
versions of µsoma; but unfortunately I've failed. There's the problem of
interjecting between fixpoint updates on keyboard or any other kind of
event. I can't quite figure out how to do this properly at the moment.
All I do leads to device conflicts; which means we'd probably have to
have the `fixpoint` device cooperate with us in some way.