Upload overall progress
I have a nasty habit of forgetting to commit and this is an example of a
particularly bad episode of that!! Two months of work!!!!!! My gosh!
This branch will be slowly removing the old stuff (e.g. soma6), the
previous rack impl, and will include patches to DwUIR. It's already
largely incompatible with `kappa`. That's why I'm not sure merging it
into `kappa` eventually will be a good idea.

Currently this branch is pretty much un-buildable for anyone but me.
Again we're in a situation where there's no working "front-end" for
Wirewright. Soon enough there will be, though. I've got some very
interesting experiments indeed running here!