d7: Upload new implementation
Minor to no refinements on the outside but lots of major changes on the
inside. It's more performant now, although not by much, maybe 2x,
because we're not doing combinatorial search now (not that I see it
anywhere, maybe we do). The code is still fairly stupid but will do for
now. Hopefully I'll be able to make progress on more interesting parts
rather than this, the guttiest guts of the engine.