d7: Upload recent work on D7
A lot of things are still *extremely* unclear, in particular a) how to
do entanglement so that it's friends with a git-like "timeline" of some
sort without exploding the world or itself, and b) which order to run
transformations on the circuit in. Should it be post-`D7.step`,
pre-step? What about observation (e.g. termspaces observing sensor and
appearance complexes in the circuit)? How should subcircuits be handled
in either case? These questions are a severe headache for me -- I have
no idea, and I don't even know if the answers would matter.

On performance, I think we're going to be fine. The solver where the
heavy stuff happens isn't the hot part even in profiles of the current
prototype-ish implementation. The honor is fold()'s and its friends'.
They take a lot of time (mainly in M1 and Dict, and both require the
dict optimizations which are scheduled but low-priority right now).