term: Remove `ITerm#to_s` in favor of instance-specific `to_s`
Right now all of them point to `inspect` as before, as they probably
should. Ideally, in the future, all of the inspect's will in turn point
to ML.compact, since it is the representation that is responsible for
printing the term and not the term itself. At least that makes sense to
me, anything else currently doesn't.