term: Remove `Term#upcast`, `ITerm#upcast` in favor of `Term.of`
At this point we're left with only autocast and downcast in `ITerm`.
`downcast` could probably be moved to `Term.dncast` so that we don't
invade term instances. Then we'd be warring autocast. It's an implicit
mess spread evenly throughout the code, I don't know if it's going to be
easy to refactor it into something sensible, but I'll try... Autocast is
one of those things you'd want to design perfectly at the get go; any
worse, and the code starts to smell in a very bad way.