d7: Upload recent work on D7
More experimentation. Almost have dwuir.in.wwml running properly (the
engine remains uncommitted since it's in heavy flux). The only remaining
obstacle is an `image` node and some sort of way to write files to disk.
We can already watch paths (again, the engine is uncommitted but will
possibly be committed at some point in the future). But creating them is
another story.

Pretty soon I'll have to refactor M1 though because there's a lot of
"technical debt" there, especially with backmaps, and I'm not sure I
want to commit to the current architecture too much. Rewriters aren't
there (currently they're in baz5 but that's not what I want). Pattern
grammars are missing. The backmap engine is shitty and slow. M1 itself
is cryptic and is missing or mis-implementing some operators; I barely
remember how M1 works, the part of it that's still in pattern7 (its good
part is in src/ and I understand it perfectly well).