d7: Upload recent D7 work and associated DwUIR changes
The current implementation is starting to fall apart due to
implementation complexity and lack of composition for program walks. The
regime stuff on the other hand is working very well.

In the next, 9th revision of D7, I think we'll keep regime stuff, remove
reaction and chats/events (I think they're extraneous). For errors we'll
be using node pairspart, with `msg: (err "...")`. This should make it
very easy to display such errors in a speech bubble or something like
that later on. It would be nice to have cache storage associated with
devices, especially once we have rewriters. Alternatively, we could
cache rewriters outside of D7, just like we cache patterns. The only
problem right now is we don't have a good cache implementation so we'd
risk leaking lots and lots of memory.

Everything else is more or less clear, I suppose. The engine prototype
I'm working on seems to have grown into an architecture that *could*
work, if refined.

One more thing with the engine is that again, walks must be composable,
meaning, here in particular, that they should return a patch list and
never patch stuff themselves.

This is tightly linked to termspaces and other global-ish passes, which
we should banish. Bottom-up BFS is the only way, it seems, something
like IDDFS or what we do with backmaps. Patch as we escape nested
circuits and toplevel circuit. Global state should be stored aside and
updated as we ascend. I think this *could* work. Maybe. Just maybe...