term: Fixup to `Term.case` reorganization
Unfortunately Crystal emits malformed machine code for it (for whatever
reason; it seems to be related to the use of the block in `scan` and how
its return type is determined). We'll have to use iterators for now.
It's much more expensive in terms of memory; unfortunately, Crystal
doesn't have escape analysis so they go to GC heap. But until we have
extremely slim dicts, anything works. Even a single env dict node is
larger than a whole chain of iterators.