ml: Upload an unexpected rewrite/redesign of WwML reader
I didn't plan for this but it just happened. (As it always does! That's
why my plan is more than 500 lines already with little to no progress on
anything in particular.)

While redoing `Term::Sym` stuff I've noticed the nasty un-printable byte
start and end, rule ids, all that crappy crap. Unfortunately to remove
it I had to do the reader properly. Now at least it's a bit easier to
work with, the reader, I mean, although it did gain some weight (as in
complexity and LOC, maybe 1.5kloc more?)

- Reorganize ml/. Now things are named the way they should (more or
less, you can't be perfect with that). Lexer is lexer, etc. No more
pointless fanciness with names, as in lexeme/reader.
- We now have a proper parser! We're emitting AST! It's heavily
union-type'd (compared to the way one is usually designing ASTs), and is
found in tree/. Then there's `ML::Renderer` which is a generic module
with about 50 (maybe less) overloads of `render`. It converts the AST to
terms.
- We now have proper, robust (well, I hope so!) srcmap tracking. Not
just certain kinds of dicts, but everything, dicts or not. Even doc
comments on rules are properly tracked now!
- The last time I benchmarked it the new reader was ~2.1x faster than
the old one. However, since that time I've added all the `Tsrc` business
so I'd not be surprised if it turns out slower in the end, although
probably it won't.
- I've removed the wonky `↢` and `↣` (readtime append and prepend)
syntax because it's just extraneous, nobody would need that, not even
me.
- New shorthands include `^:x` -> `x: ^x`, `@:x` -> `x: @x`, selector
`@x_` -> `x: @x_`, all-items `⟨⊚ a b c⟩` -> `(%all ⟨a⟩ ⟨b⟩ ⟨c⟩)`,
all-leaves `⟪a b c⟫` -> `(%all (%leaf a) (%leaf b) (%leaf c))`. These
forms I'm seeing more and more, especially while working on D7/Rack.

Unfortunately we're quickly running out of Unicode brackets. The ones
that remain are barely distinguishable from what we already have, as if
what we already have is distinguishable! A nasty position to be in, to
be honest.