term: Upload the new implementation of `Term::Sym`
I'm not sure if we're *actually* faster than the previous impl in
practice, but on some benchmarks, we are. It's very hard to compete with
the previous impl because it e.g. parsed in ~25ns on my machine, and the
new one parses in ~15-17ns, as far as I remember. There's little you can
do to get faster from the Crystal side, and I'm obviously not the one to
go into C or Assembly -- I have far better & more interesting things to
do.

Other things are expected to be faster as well, especially because we're
inlining short names into the bits. Something like checking whether a
symbol is a variable in Alloy (e.g. `^x`) will not hit memory now (well,
it shouldn't, but the Alloy side of the question is maybe *a little bit*
on the shitty end). In multi-threaded scenarios, this also means little
to no fighting over who gets to read or modify the symbol table.

Regardless, symbols are a fairly insignificant part of the current
corpus of code on which I measure "practical improvement". There's no to
little MT. The majority of what the current corpus does (e.g. tests) in
terms of time is compiling patterns, as stupid as it sounds. This is a
problem in an entirely different dimension vs. symbols. It's just that
the previous impl of symbols was dirty, and this one is less dirty, and
hopefully just a tiny bit faster, or at least not too much slower.