editR: Upload progress on the new implementation of editR
- Moved to Alloy backmap applier.
- It's significantly less shittier now, and it's TDDd (to an extent), so
we have moderate to high test coverage.
- Selection handling is much less headache-y. The repr is still bad, and
there's a huge comment at the top of the codex related to that. We're
now using an explicit anchor integer instead of three-part zipper. We
convert to three-part zipper when necessary (e.g. on insertion).
Previously we did the opposite: we converted to a range-based repr from
a three-pard zipper.
- It's also very slow! Takes ~12s to run in debug-mode for 150 tests.
Release-mode is ~800ms on my machine. That it is slow is actually quite
nice, because it can be used as a benchmark for future opts. Although
the majority of time is spent in M1 which isn't exactly full of
low-hanging fruits anymore... We're mainly doing lots of useless
traversal (and absR on top of it!) so that's a nightmare even for tiny
terms. No idea how to fix that right now. Hopefully brute force
optimization will save us but if it won't I'll have to think really
hard!
- Currently it doesn't interface with the rest of the code.
- Missing feature: prompt. It needs to be rethought from scratch because
the old impl was odd.
- Missing some minor commands but I don't remember using them anyway.