dwuir: Fix newlines not being highlighted as selected in `Textual`
The whole thing must be rewritten, though. I mean the entirety of DwUIR.
When I was writing it I didn't know what I was doing, now I sort of do.
Lots of inefficiencies there as well. We're also missing one big
feature, which is auto-pan on viewports. Viewports should be able to act
as "cameras" and follow specific objects or selections. It's all a big
task though. I have other priorities at the moment.