alloy: Refactor `Alloy.compose`, remove compose cache
The compose cache is misplaced. During compose, the granularity is too
coarse for a cache to be useful. We'd want to cache on a deeper level.
Unfortunately even `Alloy.render` itself seems too coarse and too
entangled with the vars dict for caching to be useful. Alloy is pure,
though, so it must be possible. With the eval and refine procs, however,
we are even more entangled, esp. because clients like `Alloy.compose`
generate them as they go, and they're part of the cache key, so any
render cache would be useless too.