primitives: Work on primitives, cleanup
I still haven't decided what the naming style should be: should we use
verbs or nouns, for example? I like nouns but they read weird, and often
we have to noun-ify verbs, because it's quite hard to name *results* of
things, and the names that I end up using feel overly verbose or
outright quirky! The current primitives business also can't handle more
elaborate expressions (those with more "syntax" so to speak), because
they're going to be immediately evaluated and we wouldn't want that.