m1: Sort `%entries` (`AllIsolated`) output using `Term.compare`
The AllIsolated vs. All distinction is going to go but it's currently
here so we have to deal with it somehow. We don't sort on All because
there's no test for that and I don't plan keeping the code anyway, so
it's not worth the bother.

This change lets us remove `⊆` from testtool.